camera_app/requirements.txt
argcomplete==3.5.0
blinker==1.8.2
boto3==1.35.14
botocore==1.35.14
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
durationpy==0.7
Flask==3.0.3
geographiclib==2.0
geopy==2.4.1
hjson==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
pillow==10.4.0
placebo==0.9.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
setuptools==74.1.2
six==1.16.0
SQLAlchemy==2.0.34
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.5
troposphere==4.8.1
typing_extensions==4.12.2
typish==1.9.3
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
zappa==0.59.0
